Move getCheckoutEventHandler to own file
